+++ backend/src/main/java/com/study/deposit/global/security/SecurityConfig.java
@@ -1,12 +1,10 @@
-package com.study.deposit.global.auth;
+package com.study.deposit.global.security;
 
-import com.study.deposit.global.auth.CustomOAuth2UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
